+++ aqt_connector/__init__.py
@@ -1,5 +1,6 @@
 from aqt_connector._application.authentication import get_access_token as get_access_token
 from aqt_connector._application.authentication import log_in as log_in
+from aqt_connector._application.jobs import fetch_job_state as fetch_job_state
 from aqt_connector._arnica_app import ArnicaApp as ArnicaApp
 from aqt_connector._sdk_config import ArnicaConfig as ArnicaConfig
 
@@ -7,5 +8,6 @@ __all__ = [
     "ArnicaApp",
     "get_access_token",
     "log_in",
+    "fetch_job_state",
     "ArnicaConfig",
 ]