+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Changelog
 
 ## unreleased
+* Add fetch job state #8
+* Add Arnica API schemas and update the models #4
 
 ## aqt-connector 0.1.0
 * Add license and changelog. Update pyproject.toml #3